+++ src/Applications/use_case/ReplyUseCase.js
@@ -1,36 +1,31 @@
 const Reply = require('../../Domains/replies/entities/Reply');
 
 class ReplyUseCase {
-    constructor({replyRepository}) {
+    constructor({
+                    threadRepository,
+                    commentRepository,
+                    replyRepository
+                }) {
+        this._threadRepository = threadRepository;
+        this._commentRepository = commentRepository;
         this._replyRepository = replyRepository;
     }
 
-    addReply(owner, threadId, commentId, content) {
+    async addReply(owner, threadId, commentId, content) {
         const reply = new Reply({owner, threadId, commentId, content});
+        await this._threadRepository.getThreadById(threadId);
+        await this._commentRepository.getComment(threadId, commentId);
         return this._replyRepository.addReply(reply);
     }
 
-    async getReply(threadId, commentId, replyId) {
+    async deleteReply(threadId, commentId, replyId, userId) {
         this._validateString(threadId);
         this._validateString(commentId);
         this._validateString(replyId);
-        return this._replyRepository.getReply(threadId, commentId, replyId);
-    }
-
-    async getRepliesInCommentsThread(threadId, commentId) {
-        this._validateString(threadId);
-        this._validateString(commentId);
-        return this._replyRepository.getRepliesInCommentsThread(threadId, commentId);
-    }
-
-    async verifyReplyOwner(replyId, userId) {
-        this._validateString(replyId)
-        this._validateString(userId)
-        return this._replyRepository.verifyReplyOwner(replyId, userId)
-    }
-
-    async deleteReply(replyId) {
-        this._validateString(replyId);
+        await this._threadRepository.getThreadById(threadId);
+        await this._commentRepository.getComment(threadId, commentId);
+        await this._replyRepository.getReply(threadId, commentId, replyId,);
+        await this._replyRepository.verifyReplyOwner(replyId, userId)
         return this._replyRepository.deleteReply(replyId);
     }
